--- artemis-protocols/artemis-mqtt-protocol/src/test/resources/embedded-artemis-server-mqtt.xml
@@ -1,34 +1,0 @@
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one or more
-  contributor license agreements.  See the NOTICE file distributed with
-  this work for additional information regarding copyright ownership.
-  The ASF licenses this file to You under the Apache License, Version 2.0
-  (the "License"); you may not use this file except in compliance with
-  the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<configuration xmlns="urn:activemq"
-               xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-               xsi:schemaLocation="
-               urn:activemq /schema/artemis-server.xsd
-               urn:activemq:core /schema/artemis-configuration.xsd">
-   <core xmlns="urn:activemq:core">
-      <persistence-enabled>false</persistence-enabled>
-      <security-enabled>false</security-enabled>
-
-      <!-- Acceptors -->
-      <acceptors>
-         <!-- In VM acceptor -->
-         <acceptor name="in-vm">vm://0</acceptor>
-         <acceptor name="mqtt">tcp://127.0.0.1:1883?protocols=MQTT&amp;useEpoll=false&amp;useKQueue=false</acceptor>
-      </acceptors>
-
-   </core>
-</configuration>
